Add the map UI to show the route
The response object from hitting the directions endpoint directly seems
to be different than the one returned from the JS directions service --
different enough so as to be invalid according to the directions
renderer. So for now we make two requests for directions: one on the
backend to get the steps and one on the frontend to show the route in
the map.